Refactor, as_3d moved to np2tensor/nptensor
