+++ backend/src/nodes/nodes/onnx/upscale_image.py
@@ -11,11 +11,12 @@ from ...node_factory import NodeFactory
 from ...properties.inputs import OnnxModelInput, ImageInput, TileSizeDropdown
 from ...properties.outputs import ImageOutput
 from ...utils.auto_split_tiles import parse_tile_size_input, TileSize
+from ...utils.convenient_upscale import convenient_upscale
 from ...utils.auto_split import ExactTileSize
 from ...utils.onnx_auto_split import onnx_auto_split
 from ...utils.onnx_model import OnnxModel
 from ...utils.onnx_session import get_onnx_session
-from ...utils.utils import get_h_w_c, convenient_upscale
+from ...utils.utils import get_h_w_c
 from ...utils.exec_options import get_execution_options
