+++ backend/src/nodes/utils/utils.py
@@ -2,26 +2,13 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Callable, List, Tuple, Type, Union
+from typing import List, Tuple, Union
 import re
 import os
-import cv2
 
 import numpy as np
 from sanic.log import logger
 
-MAX_VALUES_BY_DTYPE = {
-    np.dtype("int8"): 127,
-    np.dtype("uint8"): 255,
-    np.dtype("int16"): 32767,
-    np.dtype("uint16"): 65535,
-    np.dtype("int32"): 2147483647,
-    np.dtype("uint32"): 4294967295,
-    np.dtype("int64"): 9223372036854775807,
-    np.dtype("uint64"): 18446744073709551615,
-    np.dtype("float32"): 1.0,
-    np.dtype("float64"): 1.0,
-}
 NUMBERS = re.compile(r"(\d+)")
 
 ALPHABET = [*"ABCDEFGHIJKLMNOPQRSTUVWXYZ"]
@@ -44,255 +31,6 @@ def get_h_w_c(image: np.ndarray) -> Tuple[int, int, int]:
     h, w = image.shape[:2]
     c = 1 if image.ndim == 2 else image.shape[2]
     return h, w, c
-
-
-def as_3d(img: np.ndarray) -> np.ndarray:
-    """Given a grayscale image, this returns an image with 3 dimensions (image.ndim == 3)."""
-    if img.ndim == 2:
-        return np.expand_dims(img.copy(), axis=2)
-    return img
-
-
-def denorm(x: np.ndarray, min_max: Tuple[float, float] = (-1.0, 1.0)) -> np.ndarray:
-    """Denormalize from [-1,1] range to [0,1]
-    formula: xi' = (xi - mu)/sigma
-    Example: "out = (x + 1.0) / 2.0" for denorm
-        range (-1,1) to (0,1)
-    for use with proper act in Generator output (ie. tanh)
-    """
-    out = (x - min_max[0]) / (min_max[1] - min_max[0])
-    return np.clip(out, 0, 1)
-
-
-def norm(x: np.ndarray) -> np.ndarray:
-    """Normalize (z-norm) from [0,1] range to [-1,1]"""
-    out = (x - 0.5) * 2.0
-    return np.clip(out, -1, 1)
-
-
-def np_bgr_to_rgb(image: np.ndarray) -> np.ndarray:
-    out: np.ndarray = image[::-1, ...]
-    return out
-
-
-def np_rgb_to_bgr(image: np.ndarray) -> np.ndarray:
-    # same operation as bgr_to_rgb(), flip image channels
-    return np_bgr_to_rgb(image)
-
-
-def np_bgra_to_rgba(image: np.ndarray) -> np.ndarray:
-    out: np.ndarray = image[[2, 1, 0, 3], ...]  # type: ignore
-    return out
-
-
-def np_rgba_to_bgra(image: np.ndarray) -> np.ndarray:
-    # same operation as bgra_to_rgba(), flip image channels
-    return np_bgra_to_rgba(image)
-
-
-def np2nptensor(
-    img: np.ndarray,
-    bgr2rgb=True,
-    data_range=1.0,  # pylint: disable=unused-argument
-    normalize=False,
-    change_range=True,
-    add_batch=True,
-) -> np.ndarray:
-    """Converts a numpy image array into a numpy Tensor array.
-    Parameters:
-        img (numpy array): the input image numpy array
-        add_batch (bool): choose if new tensor needs batch dimension added
-    """
-    if not isinstance(img, np.ndarray):  # images expected to be uint8 -> 255
-        raise TypeError("Got unexpected object type, expected np.ndarray")
-    # check how many channels the image has, then condition. ie. RGB, RGBA, Gray
-    # if bgr2rgb:
-    #     img = img[
-    #         :, :, [2, 1, 0]
-    #     ]  # BGR to RGB -> in numpy, if using OpenCV, else not needed. Only if image has colors.
-    if change_range:
-        dtype = img.dtype
-        maxval = MAX_VALUES_BY_DTYPE.get(dtype, 1.0)
-        t_dtype = np.dtype("float32")
-        img = img.astype(t_dtype) / maxval  # ie: uint8 = /255
-    # "HWC to CHW" and "numpy to tensor"
-    img = np.ascontiguousarray(np.transpose(img, (2, 0, 1))).astype(np.float32)
-    if bgr2rgb:
-        # BGR to RGB -> in tensor, if using OpenCV, else not needed. Only if image has colors.)
-        if (
-            img.shape[0] % 3 == 0
-        ):  # RGB or MultixRGB (3xRGB, 5xRGB, etc. For video tensors.)
-            img = np_bgr_to_rgb(img)
-        elif img.shape[0] == 4:  # RGBA
-            img = np_bgra_to_rgba(img)
-    if add_batch:
-        img = np.expand_dims(
-            img, axis=0
-        )  # Add fake batch dimension = 1 . squeeze() will remove the dimensions of size 1
-    if normalize:
-        img = norm(img)
-    return img
-
-
-def nptensor2np(
-    img: np.ndarray,
-    rgb2bgr=True,
-    remove_batch=True,
-    data_range=255,
-    denormalize=False,
-    change_range=True,
-    imtype: Type = np.uint8,
-) -> np.ndarray:
-    """Converts a Tensor array into a numpy image array.
-    Parameters:
-        img (tensor): the input image tensor array
-            4D(B,(3/1),H,W), 3D(C,H,W), or 2D(H,W), any range, RGB channel order
-        remove_batch (bool): choose if tensor of shape BCHW needs to be squeezed
-        denormalize (bool): Used to denormalize from [-1,1] range back to [0,1]
-        imtype (type): the desired type of the converted numpy array (np.uint8
-            default)
-    Output:
-        img (np array): 3D(H,W,C) or 2D(H,W), [0,255], np.uint8 (default)
-    """
-    n_dim = img.ndim
-
-    img = img.astype(np.float32)
-
-    if n_dim in (4, 3):
-        # if n_dim == 4, has to convert to 3 dimensions
-        if n_dim == 4 and remove_batch:
-            # remove a fake batch dimension
-            img = img.squeeze(0)
-
-        if img.shape[0] == 3 and rgb2bgr:  # RGB
-            # RGB to BGR -> in tensor, if using OpenCV, else not needed. Only if image has colors.
-            img_np = np_rgb_to_bgr(img)
-        elif img.shape[0] == 4 and rgb2bgr:  # RGBA
-            # RGBA to BGRA -> in tensor, if using OpenCV, else not needed. Only if image has colors.
-            img_np = np_rgba_to_bgra(img)
-        else:
-            img_np = img
-        img_np = np.transpose(img_np, (1, 2, 0))  # CHW to HWC
-    elif n_dim == 2:
-        img_np = img
-    else:
-        raise TypeError(
-            f"Only support 4D, 3D and 2D tensor. But received with dimension: {n_dim:d}"
-        )
-
-    # if rgb2bgr:
-    # img_np = img_np[[2, 1, 0], :, :] #RGB to BGR -> in numpy, if using OpenCV, else not needed. Only if image has colors.
-    # TODO: Check: could denormalize in the begining in tensor form instead
-    if denormalize:
-        img_np = denorm(img_np)  # denormalize if needed
-    if change_range:
-        img_np = np.clip(
-            data_range * img_np, 0, data_range
-        ).round()  # np.clip to the data_range
-
-    # has to be in range (0,255) before changing to np.uint8, else np.float32
-    return img_np.astype(imtype)
-
-
-def clipped(upscale: Callable[[np.ndarray], np.ndarray]) -> Callable:
-    return lambda i: np.clip(upscale(i), 0, 1)
-
-
-def ensure_3d(upscale: Callable[[np.ndarray], np.ndarray]) -> Callable:
-    """Ensures that upscale methods receive 3D input."""
-    return lambda i: upscale(as_3d(i))
-
-
-def to_target_channels(img: np.ndarray, target: int) -> np.ndarray:
-    """Adjusts the given image to have `target` number of channels."""
-    c = get_h_w_c(img)[2]
-
-    if c == target:
-        return img
-
-    if c == 1:
-        if target == 3:
-            return cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
-        if target == 4:
-            return cv2.cvtColor(img, cv2.COLOR_GRAY2BGRA)
-
-    if c == 3:
-        if target == 1:
-            return cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-        if target == 4:
-            return cv2.cvtColor(img, cv2.COLOR_BGR2BGRA)
-
-    if c == 4:
-        if target == 1:
-            return cv2.cvtColor(img, cv2.COLOR_BGRA2GRAY)
-        if target == 3:
-            return cv2.cvtColor(img, cv2.COLOR_BGRA2BGR)
-
-    raise ValueError(f"Unable to convert {c} channel image to {target} channel image")
-
-
-def with_black_and_white_backgrounds(img: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
-    c = get_h_w_c(img)[2]
-    assert c == 4
-
-    black = np.copy(img[:, :, :3])
-    white = np.copy(img[:, :, :3])
-    for c in range(3):
-        black[:, :, c] *= img[:, :, 3]
-        white[:, :, c] = (white[:, :, c] - 1) * img[:, :, 3] + 1
-
-    return black, white
-
-
-def convenient_upscale(
-    img: np.ndarray,
-    model_in_nc: int,
-    model_out_nc: int,
-    upscale: Callable[[np.ndarray], np.ndarray],
-) -> np.ndarray:
-    """
-    Upscales the given image in an intuitive/convenient way.
-
-    This method guarantees that the `upscale` function will be called with an image with
-    `model_in_nc` number of channels.
-
-    Additionally, guarantees that the number of channels of the output image will match
-    that of the input image in cases where `model_in_nc` == `model_out_nc`, and match
-    `model_out_nc` otherwise.
-    """
-    in_img_c = get_h_w_c(img)[2]
-
-    upscale = clipped(ensure_3d(upscale))
-
-    if model_in_nc != model_out_nc:
-        return upscale(to_target_channels(img, model_in_nc))
-
-    if in_img_c == model_in_nc:
-        return upscale(img)
-
-    if in_img_c == 4:
-        # Ignore alpha if single-color or not being replaced
-        unique = np.unique(img[:, :, 3])
-        if len(unique) == 1:
-            rgb = to_target_channels(
-                upscale(to_target_channels(img[:, :, :3], model_in_nc)), 3
-            )
-            unique_alpha = np.full(rgb.shape[:-1], unique[0], np.float32)
-            return np.dstack((rgb, unique_alpha))
-
-        # Transparency hack (white/black background difference alpha)
-        black, white = with_black_and_white_backgrounds(img)
-        black_up = to_target_channels(
-            upscale(to_target_channels(black, model_in_nc)), 3
-        )
-        white_up = to_target_channels(
-            upscale(to_target_channels(white, model_in_nc)), 3
-        )
-
-        alpha = 1 - np.mean(white_up - black_up, axis=2)
-        return np.dstack((black_up, alpha))
-
-    return to_target_channels(upscale(to_target_channels(img, model_in_nc)), in_img_c)
 
 
 def resize_to_side_conditional(
